+++ test/e2e-common.sh
@@ -89,8 +89,6 @@ function knative_setup() {
   enable_sugar || fail_test "Could not enable Sugar Controller Injection"
 
   unleash_duck || fail_test "Could not unleash the chaos duck"
-
-  install_feature_cm || fail_test "Could not install features configmap"
 }
 
 function scale_controlplane() {
@@ -275,14 +273,6 @@ function unleash_duck() {
     sed "s/namespace: ${KNATIVE_DEFAULT_NAMESPACE}/namespace: ${SYSTEM_NAMESPACE}/g" | \
     ko apply "${KO_FLAGS}" -f - || return $?
     if (( SCALE_CHAOSDUCK_TO_ZERO )); then kubectl -n "${SYSTEM_NAMESPACE}" scale deployment/chaosduck --replicas=0; fi
-}
-
-function install_feature_cm() {
-  KO_FLAGS="${KO_FLAGS:-}"
-  echo "install feature configmap"
-  cat test/config/config-features.yaml | \
-  sed "s/namespace: ${KNATIVE_DEFAULT_NAMESPACE}/namespace: ${SYSTEM_NAMESPACE}/g" | \
-  ko apply "${KO_FLAGS}" -f - || return $?
 }
 
 # Teardown the Knative environment after tests finish.
