+++ test/experimental/config/features.yaml
@@ -26,3 +26,4 @@ data:
   delivery-timeout: "enabled"
   new-trigger-filters: "enabled"
   eventtype-auto-create: "enabled"
+  new-apiserversource-filters: "enabled"
