+++ test/rekt/apiserversource_test.go
@@ -212,17 +212,3 @@ func TestApiServerSourceDeployment(t *testing.T) {
 
 	env.Test(ctx, t, apiserversourcefeatures.DeployAPIServerSourceWithNodeSelector())
 }
-
-func TestApiServerSourceNewFiltersFeature(t *testing.T) {
-	t.Parallel()
-
-	ctx, env := global.Environment(
-		knative.WithKnativeNamespace(system.Namespace()),
-		knative.WithLoggingConfig,
-		knative.WithTracingConfig,
-		k8s.WithEventListener,
-		environment.Managed(t),
-	)
-
-	env.TestSet(ctx, t, apiserversourcefeatures.NewFiltersFeature())
-}
